+++ Wireframe.module.php
@@ -14,7 +14,7 @@ namespace ProcessWire;
  * @method static string|Page|NullPage page($source, $args = []) Static getter (factory) method for Pages.
  * @method static string|null partial(string $partial_name, array $args = []) Static getter (factory) method for Partials.
  *
- * @version 0.12.0
+ * @version 0.13.0
  * @author Teppo Koivula <teppo@wireframe-framework.com>
  * @license Mozilla Public License v2.0 https://mozilla.org/MPL/2.0/
  */
@@ -204,7 +204,7 @@ class Wireframe extends WireData implements Module, ConfigurableModule {
      */
     public static function isInitialized(int $instanceID = null): bool {
         return \in_array(
-            \is_null($instanceID) ? wire()->instanceID : $instanceID,
+            $instanceID === null ? wire()->instanceID : $instanceID,
             static::$initialized
         );
     }
@@ -385,7 +385,7 @@ class Wireframe extends WireData implements Module, ConfigurableModule {
      */
     public function setRenderer($renderer, array $settings = []): Wireframe {
         $needs_init = !empty($settings);
-        if (\is_null($renderer)) {
+        if ($renderer === null) {
             $this->renderer = null;
             if ($this->view) $this->view->setRenderer(null);
         } else if (\is_string($renderer)) {
@@ -488,6 +488,22 @@ class Wireframe extends WireData implements Module, ConfigurableModule {
      */
     protected function addHooks() {
 
+        // make View properties directly available in TemplateFile (primarily for field rendering)
+        $this->addHookBefore('TemplateFile::render', function(HookEvent $event) {
+            $view = $event->object->view;
+            if (!$view instanceof \Wireframe\View || $event->object instanceof \Wireframe\View) {
+                // View doesn't exist or we're currently rendering a Wireframe View
+                return;
+            }
+            $event->object->setArray(array_merge($view->data(), array_filter([
+                'page' => $event->object->page, // used by field rendering
+                'value' => $event->object->value, // used by field rendering
+                'field' => $event->object->field, // used by field rendering
+                'partials' => $view->partials,
+                'placeholders' => $view->placeholders,
+            ])));
+        });
+
         // helper methods for getting or setting page layout
         $this->addHookMethod('Page::layout', $this, 'pageLayout');
         $this->addHookMethod('Page::getLayout', $this, 'pageLayout');
@@ -514,7 +530,7 @@ class Wireframe extends WireData implements Module, ConfigurableModule {
      * Look for redirect fields within config settings. If present, check if the page has a value in one of those and
      * if a redirect should be performed.
      */
-    public function ___checkRedirects() {
+    protected function ___checkRedirects() {
 
         // redirect fields from Wireframe runtime configuration
         $redirect_fields = $this->config['redirect_fields'] ?? null;
@@ -533,7 +549,7 @@ class Wireframe extends WireData implements Module, ConfigurableModule {
             // get URL from a page field
             $url = $page->get($field);
             if ($url instanceof WireArray) {
-                $url = $url->count() ? $url->first() : null;
+                $url = $url->first();
             }
             if (empty($url)) continue;
 
@@ -563,7 +579,7 @@ class Wireframe extends WireData implements Module, ConfigurableModule {
      * @param string $url Redirect URL.
      * @param bool $permanent Is this a permanent (301) redirect?
      */
-    public function ___redirect(string $url, bool $permanent) {
+    protected function ___redirect(string $url, bool $permanent) {
         $this->wire('session')->redirect($url, $permanent);
     }
 
@@ -576,24 +592,21 @@ class Wireframe extends WireData implements Module, ConfigurableModule {
      *
      * @throws WireException if no valid Page has been defined.
      */
-    public function ___initView(): \Wireframe\View {
+    protected function ___initView(): \Wireframe\View {
 
         // params
-        $page = $this->page;
-        $paths = $this->paths;
-        $ext = $this->ext;
-        $data = $this->data;
+        $page_layout = $this->page->getLayout();
 
         // initialize the View object (note: view file is set in the Wireframe::___setView() method)
         $view = new \Wireframe\View;
-        $view->setLayout($page->getLayout() === null ? 'default' : $page->getLayout());
-        $view->setTemplate($page->getViewTemplate());
-        $view->setViewsPath($paths->views);
-        $view->setLayoutsPath($paths->layouts);
-        $view->setExt($ext);
+        $view->setLayout($page_layout === null ? 'default' : $page_layout);
+        $view->setTemplate($this->page->getViewTemplate());
+        $view->setViewsPath($this->paths->views);
+        $view->setLayoutsPath($this->paths->layouts);
+        $view->setExt($this->ext);
         $view->setPage($this->page);
-        $view->setData($data);
-        $view->setPartials($this->findPartials($paths->partials));
+        $view->setData($this->data);
+        $view->setPartials($this->findPartials($this->paths->partials));
         $view->setPlaceholders(new \Wireframe\ViewPlaceholders($view));
         $view->setRenderer($this->renderer);
         $this->view = $view;
@@ -612,7 +625,7 @@ class Wireframe extends WireData implements Module, ConfigurableModule {
      *
      * @return \Wireframe\Controller|null Controller object or null.
      */
-    public function ___initController(): ?\Wireframe\Controller {
+    protected function ___initController(): ?\Wireframe\Controller {
 
         // get a new Controller instance and store it both locally, as a run-time property of current Page object, and
         // as a reference within the View object
@@ -635,7 +648,7 @@ class Wireframe extends WireData implements Module, ConfigurableModule {
     public function ___setView(?string $view = null): Wireframe {
 
         // first check if a predefined view name was provided
-        if (!\is_null($view)) {
+        if ($view !== null) {
             $view = basename($view);
             $this->view->setView($view);
             $this->page->setView($view);
@@ -746,7 +759,7 @@ class Wireframe extends WireData implements Module, ConfigurableModule {
             $this->settings_hash,
             empty($data) ? '' : md5(json_encode($data)),
             $view->getTemplate(),
-            $controller,
+            $controller === null ? '' : \get_class($controller),
             $view->getFilename(),
             $view->getLayout(),
             $view->getExt(),
@@ -918,7 +931,7 @@ class Wireframe extends WireData implements Module, ConfigurableModule {
     protected function pageController(HookEvent $event) {
         if ($event->method == 'getController') {
             $controller = $event->object->_wireframe_controller ?: null;
-            if (\is_null($controller)) {
+            if ($controller === null) {
                 $controller = $this->getController($event->object);
             }
             $event->return = $controller;
@@ -1034,7 +1047,7 @@ class Wireframe extends WireData implements Module, ConfigurableModule {
                 if (\is_array($value) && !empty($value)) {
                     $invalid_value = false;
                     $this->setRenderer($value[0], $value[1] ?? []);
-                } else if (\is_null($value) || \is_string($value) || $value instanceof Module) {
+                } else if ($value === null || \is_string($value) || $value instanceof Module) {
                     $invalid_value = false;
                     $this->setRenderer($value);
                 }
@@ -1067,7 +1080,8 @@ class Wireframe extends WireData implements Module, ConfigurableModule {
     /**
      * Set values from an array
      *
-     * This method is a wrapper for the set() method, with support for multiple values as an associative array.
+     * This method is a wrapper for the set() method that supports multiple values as an associative array and
+     * automatically generates a hash of the settings array (for caching purposes).
      *
      * @param array $values Values as an associative array.
      * @return Wireframe Self-reference.
@@ -1110,21 +1124,25 @@ class Wireframe extends WireData implements Module, ConfigurableModule {
      *
      * @param string $path Base directory.
      * @param string|null $ext File extension (optional).
+     * @param bool $use_cache Use cache? Defaults to `true`.
      * @return \Wireframe\Partials A container populated with Partials.
      */
-    protected function findPartials(string $path, string $ext = null): \Wireframe\Partials {
-        $cache_key = 'files:' . $path . ':' . $ext;
-        $files = $this->cache[$cache_key] ?? [];
-        if (empty($files)) {
+    protected function findPartials(string $path, string $ext = null, bool $use_cache = true): \Wireframe\Partials {
+        $files = [];
+        if ($use_cache) {
+            $cache_key = 'partials:' . $path . ':' . $ext;
+            $files = $this->cache[$cache_key] ?? [];
+        }
+        if (!$use_cache || empty($files)) {
             foreach (\glob($path . '*') as $file) {
                 $name = \basename($file);
                 if (\strpos($name, ".") === 0) continue;
                 if (\is_dir($file)) {
-                    $files[$name] = $this->findPartials("{$file}/", $ext);
+                    $files[$name] = $this->findPartials("{$file}/", $ext, false);
                 } else {
                     $file_data = [];
                     $ext_pos = \strrpos($name, '.');
-                    if (\is_null($ext)) {
+                    if ($ext === null) {
                         if ($ext_pos !== false) {
                             $temp_ext = \ltrim(\substr($name, $ext_pos), '.');
                             $file_data[$temp_ext] = $file;
@@ -1144,10 +1162,12 @@ class Wireframe extends WireData implements Module, ConfigurableModule {
                 $files = new \Wireframe\Partials();
                 $files->setPath($path);
                 foreach ($temp_files as $key => $file) {
-                    $files->{$key} = \is_object($file) ? $file : new \Wireframe\Partial($file);
+                    $files->{$key} = \is_object($file) ? $file : $this->wire(new \Wireframe\Partial($file));
                 }
             }
-            $this->cache[$cache_key] = $files;
+            if ($use_cache) {
+                $this->cache[$cache_key] = $files;
+            }
         }
         return $this->wire($files);
     }
@@ -1155,12 +1175,20 @@ class Wireframe extends WireData implements Module, ConfigurableModule {
     /**
      * Get controller object
      *
-     * @param Page $page
+     * - If a Page object is provided as the first argument, this method returns a Controller for that Page
+     * - If a Page object is provided as the first argument and a template name as the second one, this method will
+     *   attempt to instantiate and return a Controller for provided Page and template name.
+     * - If neither Page nor template name are provided, this method returns Wireframe instance Controller property.
+     *
+     * @param Page|null $page
      * @param string|null $template_name
      * @return \Wireframe\Controller|null
      */
-    protected function getController(Page $page, ?string $template_name = null): ?\Wireframe\Controller {
-        if (\is_null($template_name)) {
+    public function getController(?Page $page = null, ?string $template_name = null): ?\Wireframe\Controller {
+        if ($page === null) {
+            return $this->controller;
+        }
+        if ($template_name === null) {
             if (isset($page->_wireframe_controller)) {
                 return $page->_wireframe_controller == '' ? null : $page->_wireframe_controller;
             }
@@ -1170,7 +1198,7 @@ class Wireframe extends WireData implements Module, ConfigurableModule {
         $controller_class = '\Wireframe\Controller\\' . $controller_name . 'Controller';
 
         if (class_exists($controller_class)) {
-            return new $controller_class($this->wire(), $page, $this->view);
+            return $this->wire(new $controller_class($page, $this->view));
         }
         return null;
     }